move report class to separate file
